pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["Nikodem <nikodemmlynski08@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.124.0"
uvicorn = {extras = ["standard"], version = "^0.38.0"}
python-dotenv = "^1.2.1"
psycopg2-binary = "^2.9.11"
sqlmodel = "^0.0.27"
pydantic-settings = "^2.12.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
